* fully functional landing page with the 'official' style
